% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_contributions.R
\name{aggregate_contributions}
\alias{aggregate_contributions}
\title{Build an R-universe \code{packages.json} manifest from a collection of
text file contributions with individual package URLs.}
\usage{
aggregate_contributions(
  input = getwd(),
  output = "packages.json",
  owner_exceptions = character(0L)
)
}
\arguments{
\item{input}{Character of length 1, directory path with the
text file listings of R releases.}

\item{output}{Character of length 1, file path where the
R-universe \code{packages.json} file will be written.}

\item{owner_exceptions}{Character vector of URLs of GitHub owners
where \code{"branch": "*release"} should be omitted. Example:
\code{"https://github.com/cran"}.}
}
\value{
NULL (invisibly)
}
\description{
Build an R-universe \code{packages.json} manifest from a collection
of text file contributions with individual package URLs.
}
\keyword{internal}
