% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_cran_url.R
\name{assert_cran_url}
\alias{assert_cran_url}
\title{Validate the URL of a package on CRAN.}
\usage{
assert_cran_url(name, url)
}
\arguments{
\item{name}{Character of length 1, contributed name of the package.}

\item{url}{Character of length 1, contributed URL of the package.}
}
\value{
A character string if there is a problem with the package entry,
otherwise \code{NULL} if there are no issues.
}
\description{
If the package is on CRAN, validate that the contributed URL
appears in the CRAN-hosted DESCRIPTION file.
}
\keyword{internal}
