% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_package_description.R
\name{assert_package_description}
\alias{assert_package_description}
\title{Assert package \code{DESCRIPTION}}
\usage{
assert_package_description(name, url)
}
\arguments{
\item{name}{Character string, name of the package listing contribution
to R-multiverse.}

\item{url}{Character string, URL of the package on GitHub or GitLab.}
}
\value{
A character string with an informative message if a problem
was found. Otherwise, \code{NULL} if there are no issues.
}
\description{
Run basic assertions on the \code{DESCRIPTION} file of a package
on GitHub or GitLab.
}
\keyword{internal}
