% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_package.R
\name{assert_package_url}
\alias{assert_package_url}
\title{Validate a Package URL}
\usage{
assert_package_url(name, url)
}
\value{
A character string if there is a problem with the URL for the given
package name, otherwise \code{NULL} if there are no issues.
}
\description{
Validate that the package URL is in the description file if on
CRAN.
}
\keyword{internal}
