% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_release_exists.R
\name{assert_release_exists}
\alias{assert_release_exists}
\title{Check for a release.}
\usage{
assert_release_exists(url)
}
\arguments{
\item{url}{Either a character string with the package URL or
a custom JSON string with a package entry.}
}
\value{
A character string if there is a problem with the package entry,
otherwise \code{NULL} if there are no issues.
}
\description{
Check for a release.
}
\keyword{internal}
