% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_checks.R
\name{check_checks}
\alias{check_checks}
\title{Check R-universe package check results.}
\usage{
check_checks(repo = "https://multiverse.r-multiverse.org", index = NULL)
}
\arguments{
\item{repo}{Character of length 1, URL of the repository
of R package candidates for production.}

\item{index}{For testing purposes only,
an optional pre-computed data frame with details about packages.}
}
\value{
A named list of information about packages which do not comply
with \code{DESCRPTION} checks. Each name is a package name,
and each element contains specific information about
non-compliance.
}
\description{
Check R-universe package check results.
}
\details{
This function scrapes the \code{src/contrib/PACKAGES.json} file
of the universe to check the data in the \code{DESCRIPTION} files of packages
for compliance. Right now, the only field checked is \verb{Remotes:}.
A packages with a \verb{Remotes:} field in the \code{DESCRIPTION} file may
depend on development versions of other packages and so are
excluded from the production universe.
}
\section{Checks}{

Functions like \code{\link[=check_versions]{check_versions()}} and \code{\link[=check_descriptions]{check_descriptions()}}
the packages in \url{https://multiverse.r-multiverse.org}. Only packages
that pass these checks will appear in
\url{https://production.r-multiverse.org}. For a complete list of checks,
see \url{https://r-multiverse.org/multiverse.internals/reference.html#checks}
}

\examples{
  str(check_descriptions(repo = "https://multiverse.r-multiverse.org"))
}
\seealso{
Other checks: 
\code{\link{check_descriptions}()},
\code{\link{check_versions}()}
}
\concept{checks}
