% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_descriptions.R
\name{check_descriptions}
\alias{check_descriptions}
\title{Check package \code{DESCRIPTION} files.}
\usage{
check_descriptions(repo = "https://multiverse.r-multiverse.org", index = NULL)
}
\arguments{
\item{repo}{Character of length 1, URL of the repository
of R package candidates for production.}

\item{index}{For testing purposes only,
an optional pre-computed data frame with the package index.}
}
\value{
A named list of information about packages which do not comply
with \code{DESCRPTION} checks. Each name is a package name,
and each element contains specific information about
non-compliance.
}
\description{
Check the \code{DESCRIPTION} files of packages for specific
issues.
}
\details{
This function scrapes the \code{src/contrib/PACKAGES.json} file
of the universe to check the data in the \code{DESCRIPTION} files of packages
for compliance. Right now, the only field checked is \verb{Remotes:}.
A packages with a \verb{Remotes:} field in the \code{DESCRIPTION} file may
depend on development versions of other packages and so are
excluded from the production universe.
}
\examples{
  str(check_descriptions(repo = "https://multiverse.r-multiverse.org"))
}
