% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_snapshot.R
\name{date_snapshot}
\alias{date_snapshot}
\title{Snapshot date}
\usage{
date_snapshot(today = Sys.Date())
}
\arguments{
\item{today}{A \code{"Date"} object with today's date, or an object
that \code{\link[=as.Date]{as.Date()}} can convert to a \code{"Date"} object.}
}
\value{
Character string in \code{yyyy-mm-dd} format.
}
\description{
Get the target snapshot date of the current or most recent
Staging period.
}
\examples{
  date_snapshot()
}
\seealso{
Other snapshot metadata: 
\code{\link{date_staging}()},
\code{\link{r_version_staging}()}
}
\concept{snapshot metadata}
