% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_staging.R
\name{date_staging}
\alias{date_staging}
\title{Staging date}
\usage{
date_staging(today = Sys.Date())
}
\arguments{
\item{today}{A \code{"Date"} object with today's date, or an object
that \code{\link[=as.Date]{as.Date()}} can convert to a \code{"Date"} object.}
}
\value{
Character string in \code{yyyy-mm-dd} format.
}
\description{
Get the start date of the current or most recent
Staging period.
}
\examples{
  date_staging()
}
\seealso{
Other snapshot metadata: 
\code{\link{date_snapshot}()},
\code{\link{r_version_staging}()}
}
\concept{snapshot metadata}
