% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_meta.R
\name{filter_meta}
\alias{filter_meta}
\title{Filter \code{PACKAGES} and \code{PACKAGES.gz} metadata files.}
\usage{
filter_meta(path_meta, path_staging)
}
\arguments{
\item{path_meta}{Directory path where \code{PACKAGES} and \code{PACKAGES.gz} files
reside locally.}

\item{path_staging}{Path to a GitHub clone of the Staging universe.}
}
\description{
Filter the \code{PACKAGES} and \code{PACKAGES.gz} files to
only list certain packages.
}
\examples{
\dontrun{
path_meta <- tempfile()
dir.create(path_meta)
mock <- system.file(
  file.path("mock", "meta"),
  package = "multiverse.internals",
  mustWork = TRUE
)
file.copy(mock, path_meta, recursive = TRUE)
path_staging <- tempfile()
url_staging <- "https://github.com/r-multiverse/staging"
gert::git_clone(url = url_staging, path = path_staging)
filter_meta(path_meta, path_staging)
}
}
\seealso{
Other staging: 
\code{\link{freeze_dependencies}()},
\code{\link{rclone_includes}()},
\code{\link{stage_candidates}()}
}
\concept{staging}
