% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_packages.R
\name{filter_packages}
\alias{filter_packages}
\title{Filter \code{PACKAGES} and \code{PACKAGES.gz} files.}
\usage{
filter_packages(path, staged)
}
\arguments{
\item{path}{Directory path where \code{PACKAGES} and \code{PACKAGES.gz} files
reside locally.}

\item{staged}{Path to the \code{"staged.json"} file which lists staged packages
in the Staging universe.}
}
\description{
Filter the \code{PACKAGES} and \code{PACKAGES.gz} files to
only list certain packages.
}
\examples{
\dontrun{
  path <- tempfile()
  dir.create(path)
  mock <- system.file(
    "packages",
    package = "multiverse.internals",
    mustWork = TRUE
  )
  file.copy(mock, path, recursive = TRUE)
  staged <- tempfile()
  jsonlite::write_json(c("crew", "mirai"), staged, pretty = TRUE)
  filter_packages(path, staged)
}
}
\seealso{
Other staging: 
\code{\link{freeze_dependencies}()},
\code{\link{rclone_includes}()},
\code{\link{stage_candidates}()}
}
\concept{staging}
