% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_versions.R
\name{get_current_versions}
\alias{get_current_versions}
\title{Get the current versions of packages}
\usage{
get_current_versions(repos = "https://r-releases.r-universe.dev")
}
\arguments{
\item{repos}{Character string of package repositories to track.}
}
\value{
A data frame of packages with their current versions and hashes.
}
\description{
Get the current versions of packages in the repos.
}
\keyword{internal}
