% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo_file.R
\name{get_repo_file}
\alias{get_repo_file}
\title{Get repo file}
\usage{
get_repo_file(url, path)
}
\arguments{
\item{url}{Character string, URL of a GitHub or GitLab repository.}

\item{path}{Character string, path to the file in the repository.}
}
\value{
A character string with the contents of the file.
}
\description{
Get the contents of a file from a GitHub or GitLab repo.
}
\keyword{internal}
