% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_issue.R
\name{interpret_issue}
\alias{interpret_issue}
\title{Interpret a set of package issues}
\usage{
interpret_issue(path)
}
\arguments{
\item{path}{Character string, file path to a JSON issue file
of a package.}
}
\value{
A character string summarizing the issues of a package in prose.
}
\description{
Summarize the issues of a package in human-readable text.
}
\seealso{
Other issues: 
\code{\link{issues_checks}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_descriptions}()},
\code{\link{issues_versions}()}
}
\concept{issues}
