% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_issues.R
\name{interpret_issues}
\alias{interpret_issues}
\title{Interpret the issues of a package}
\usage{
interpret_issues(package, issues)
}
\arguments{
\item{package}{Character string, name of the package.}

\item{issues}{A list with one issue per package. Obtained by
reading the results of \code{\link[=record_issues]{record_issues()}}.}
}
\value{
A character string summarizing the issues of a package in prose.
}
\description{
Summarize the issues of a package in human-readable text.
}
\seealso{
Other issues: 
\code{\link{issues_checks}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_descriptions}()},
\code{\link{issues_versions}()}
}
\concept{issues}
