% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_status.R
\name{interpret_status}
\alias{interpret_status}
\title{Interpret the status of a package}
\usage{
interpret_status(package, status)
}
\arguments{
\item{package}{Character string, name of the package.}

\item{status}{A list with one status entry per package. Obtained by
reading the results of \code{\link[=record_status]{record_status()}}.}
}
\value{
A character string summarizing the status of a package in prose.
}
\description{
Summarize the status of a package in human-readable text.
}
\seealso{
Other status: 
\code{\link{status_checks}()},
\code{\link{status_dependencies}()},
\code{\link{status_descriptions}()},
\code{\link{status_versions}()},
\code{\link{update_status}()}
}
\concept{status}
