% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_member_organization.R
\name{is_member_organization}
\alias{is_member_organization}
\title{Check GitHub organization membership}
\usage{
is_member_organization(user, organizations)
}
\arguments{
\item{user}{Character string, GitHub user name.}

\item{organizations}{Character vector of names of GitHub organizations.}
}
\value{
\code{TRUE} if the user is a member of at least one of the given
GitHub organizations, \code{FALSE} otherwise.
}
\description{
Check if a GitHub user is a member of at least one
of the organizations given.
}
\keyword{internal}
