% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_advisories.R
\name{issues_advisories}
\alias{issues_advisories}
\title{Report package advisories}
\usage{
issues_advisories(meta = meta_packages())
}
\arguments{
\item{meta}{Package metadata from \code{\link[=meta_packages]{meta_packages()}}.}
}
\value{
A data frame with one row for each problematic package
and columns with details.
}
\description{
Report packages whose current versions have advisories
in the R Consortium Advisory Database:
\url{https://github.com/RConsortium/r-advisory-database}
}
\examples{
\dontrun{
issues_advisories()
}
}
\seealso{
Other issues: 
\code{\link{issues_dependencies}()},
\code{\link{issues_licenses}()},
\code{\link{issues_r_cmd_check}()},
\code{\link{issues_remotes}()},
\code{\link{issues_synchronization}()},
\code{\link{issues_version_conflicts}()},
\code{\link{issues_versions}()}
}
\concept{issues}
