% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_checks.R
\name{issues_checks}
\alias{issues_checks}
\title{Report issues from R-universe package check results.}
\usage{
issues_checks(meta = multiverse.internals::meta_checks())
}
\arguments{
\item{meta}{A data frame with R-universe package check results
returned by \code{\link[=meta_checks]{meta_checks()}}.}
}
\value{
A named list of information about packages which do not comply
with \code{DESCRPTION} checks. Each name is a package name,
and each element contains specific information about
non-compliance.
}
\description{
Check R-universe package check results.
}
\details{
\code{\link[=issues_checks]{issues_checks()}} reads output from
the R-universe check API
to scan all R-multiverse packages for issues that may have
happened during building and testing.
}
\section{Package issues}{

Functions like \code{\link[=issues_versions]{issues_versions()}} and \code{\link[=issues_descriptions]{issues_descriptions()}}
perform health checks for all packages in R-multiverse.
Only packages that pass these checks go to the production repository at
\url{https://production.r-multiverse.org}. For a complete list of checks, see
the \verb{issues_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_issues]{record_issues()}} gathers
together all the issues about R-multiverse packages.
}

\examples{
  meta <- meta_checks(repo = "https://wlandau.r-universe.dev")
  issues <- issues_checks(meta = meta)
  str(issues)
}
\seealso{
Other issues: 
\code{\link{issues_dependencies}()},
\code{\link{issues_descriptions}()},
\code{\link{issues_versions}()}
}
\concept{issues}
