% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_descriptions.R
\name{issues_descriptions}
\alias{issues_descriptions}
\title{Report \code{DESCRIPTION}-level issues.}
\usage{
issues_descriptions(meta = meta_packages())
}
\arguments{
\item{meta}{A data frame with R-universe package check results
returned by \code{\link[=meta_checks]{meta_checks()}}.}
}
\value{
A named list of information about packages which do not comply
with \code{DESCRPTION} checks. Each name is a package name,
and each element contains specific information about
non-compliance.
}
\description{
Report issues with \code{DESCRIPTION}-level metadata of packages.
}
\details{
\code{\link[=issues_descriptions]{issues_descriptions()}} reports specific issues in the
\code{DESCRIPTION}-level metadata of packages:
\enumerate{
\item Security advisories at
\url{https://github.com/RConsortium/r-advisory-database}.
\item Licenses that cannot be verified as free and open-source.
\item The presence of a \code{"Remotes"} field.
\item A lower version number on R-multiverse than on CRAN,
if the package is also published there.
}
}
\section{Package issues}{

Functions like \code{\link[=issues_versions]{issues_versions()}} and \code{\link[=issues_descriptions]{issues_descriptions()}}
perform health checks for all packages in R-multiverse.
For a complete list of checks, see
the \verb{issues_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference/index.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_issues]{record_issues()}} gathers
together all the issues about R-multiverse packages.
}

\examples{
  meta <- meta_packages(repo = "https://wlandau.r-universe.dev")
  issues <- issues_descriptions(meta = meta)
  str(issues)
}
\seealso{
Other issues: 
\code{\link{issues_checks}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_versions}()}
}
\concept{issues}
