% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_licenses.R
\name{issues_licenses}
\alias{issues_licenses}
\title{Report package license issues}
\usage{
issues_licenses(meta = meta_packages())
}
\arguments{
\item{meta}{Package metadata from \code{\link[=meta_packages]{meta_packages()}}.}
}
\description{
Report packages without standard free and open-source
licenses.
}
\examples{
\dontrun{
issues_licenses()
}
}
\seealso{
Other issues: 
\code{\link{issues_advisories}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_r_cmd_check}()},
\code{\link{issues_remotes}()},
\code{\link{issues_synchronization}()},
\code{\link{issues_version_conflicts}()},
\code{\link{issues_versions}()}
}
\concept{issues}
