% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_r_cmd_check.R
\name{issues_r_cmd_check}
\alias{issues_r_cmd_check}
\title{R-universe package \verb{R CMD check} issues.}
\usage{
issues_r_cmd_check(meta = meta_packages())
}
\arguments{
\item{meta}{Package metadata from \code{\link[=meta_packages]{meta_packages()}}.}
}
\value{
A data frame with one row for each problematic
package and columns for the package
names and \verb{R CMD check} issues.
}
\description{
Report issues from \verb{R CMD check} on R-universe.
}
\details{
\code{\link[=status_r_cmd_check]{status_r_cmd_check()}} reads output from
the R-universe \verb{R CMD check} results API
to scan all R-multiverse packages for status that may have
happened during building and testing.
}
\section{Package status}{

Functions like \code{\link[=status_versions]{status_versions()}} and \code{\link[=status_descriptions]{status_descriptions()}}
perform health checks for all packages in R-multiverse.
For a complete list of checks, see
the \verb{status_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference/index.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_status]{record_status()}} gathers
together all the status about R-multiverse packages.
}

\examples{
\dontrun{
issues_r_cmd_check()
}
}
\seealso{
Other issues: 
\code{\link{issues_advisories}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_licenses}()},
\code{\link{issues_remotes}()},
\code{\link{issues_version_conflicts}()},
\code{\link{issues_versions}()}
}
\concept{issues}
