% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_remotes.R
\name{issues_remotes}
\alias{issues_remotes}
\title{Report packages with \verb{Remotes:} fields.}
\usage{
issues_remotes(meta = meta_packages())
}
\arguments{
\item{meta}{Package metadata from \code{\link[=meta_packages]{meta_packages()}}.}
}
\description{
Report packages with \verb{Remotes:} fields in the \code{DESCRIPTION}
file.
}
\examples{
\dontrun{
issues_remotes()
}
}
\seealso{
Other issues: 
\code{\link{issues_advisories}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_licenses}()},
\code{\link{issues_r_cmd_check}()},
\code{\link{issues_version_conflicts}()},
\code{\link{issues_versions}()}
}
\concept{issues}
