% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_version_conflicts.R
\name{issues_version_conflicts}
\alias{issues_version_conflicts}
\title{Report packages with version conflicts in another repository.}
\usage{
issues_version_conflicts(meta = meta_packages(), repo = "cran")
}
\arguments{
\item{meta}{Package metadata from \code{\link[=meta_packages]{meta_packages()}}.}

\item{repo}{Character string naming the repository to compare versions.}
}
\value{
A data frame with one row for each problematic package
and columns with details.
}
\description{
Report packages with higher versions in
different repositories.
A higher version in a different repository could cause that repository
to override R-multiverse in \code{install.packages()}.
}
\examples{
\dontrun{
issues_version_conflicts()
}
}
\seealso{
Other issues: 
\code{\link{issues_advisories}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_licenses}()},
\code{\link{issues_r_cmd_check}()},
\code{\link{issues_remotes}()},
\code{\link{issues_versions}()}
}
\concept{issues}
