% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_checks.R
\name{meta_checks}
\alias{meta_checks}
\title{List metadata about R-universe package checks}
\usage{
meta_checks(repo = "https://community.r-multiverse.org")
}
\arguments{
\item{repo}{Character of length 1, URL of the package repository.
R-multiverse uses \code{"https://community.r-multiverse.org"}.}
}
\value{
A data frame with one row per package and columns with
package check results.
}
\description{
List package checks results reported by the
R-universe package API.
}
\examples{
meta_checks(repo = "https://wlandau.r-universe.dev")
}
\concept{list}
