% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{multiverse.internals-package}
\alias{multiverse.internals-package}
\title{multiverse.internals: Internal Infrastructure for R-multiverse.}
\description{
Internal Infrastructure for R-multiverse.
}
\concept{help}
