% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_snapshot.R
\name{propose_snapshot}
\alias{propose_snapshot}
\title{Propose snapshot}
\usage{
propose_snapshot(
  path_staging,
  repo_staging = "https://staging.r-multiverse.org",
  types = c("win", "mac"),
  mock = NULL
)
}
\arguments{
\item{path_staging}{Character string, directory path to the source
files of the staging universe.}

\item{repo_staging}{Character string, URL of the staging universe.}

\item{types}{Character vector, what to pass to the \code{types} field in the
snapshot API URL. Controls the types of binaries and documentation
included in the snapshot.}

\item{mock}{For testing purposes only, a named list of data frames
for inputs to various intermediate functions.}
}
\value{
\code{NULL} (invisibly). Called for its side effects.
\code{\link[=propose_snapshot]{propose_snapshot()}} writes \code{snapshot.json} with an R-universe-like
manifest of the packages recommended for the snapshot, and a
\code{snapshot.url} file containing an R-universe snapshot API URL
to download those packages. Both these files are written to the
directory given by the \code{path_staging} argument.
}
\description{
Propose a Production snapshot of Staging.
}
\details{
\code{\link[=propose_snapshot]{propose_snapshot()}} proposes a snapshot of Staging
to migrate to Production. The recommended snapshot is the list of
packages for which (1) the build and check results of the current
release are in Staging, and (2) there are no issues.
Writes \code{snapshot.json} with an R-universe-like manifest
of the packages recommended for the snapshot, and a
\code{snapshot.url} file containing an R-universe snapshot API URL
to download those packages. Both these files are written to the
directory given by the \code{path_staging} argument.
}
\examples{
\dontrun{
url_staging = "https://github.com/r-multiverse/staging"
path_staging <- tempfile()
gert::git_clone(url = url_staging, path = path_staging)
propose_snapshot(
  path_staging = path_staging,
  repo_staging = "https://staging.r-multiverse.org"
)
}
}
\seealso{
Other staging: 
\code{\link{staging_is_active}()},
\code{\link{update_staging}()}
}
\concept{staging}
