% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{r.releases.internals-package}
\alias{r.releases.internals-package}
\title{r.releases.internals: Internal infrastructure for an R universe
of package releases.}
\description{
Utilities for an R universe of package releases.
}
\concept{help}
