% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{r.releases.utils-package}
\alias{r.releases.utils-package}
\title{r.releases.utils: Utilities for an R universe of package releases.}
\description{
Utilities for an R universe of package releases..
}
\concept{help}
