% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_version_staging.R, R/utils_staging.R
\name{r_version_staging}
\alias{r_version_staging}
\title{Staging R version}
\usage{
r_version_staging()

r_version_staging()
}
\value{
A data frame with one row and columns for the \code{POSIXct} date,
patch version, and minor version of the targeted version of R.

A data frame with one row and columns for the \code{POSIXct} date,
patch version, and minor version of the targeted version of R.
}
\description{
Get the version of R from the date of the start of the
most recent Staging period.

Get the version of R from the date of the start of the
most recent Staging period.
}
\examples{
  r_version_staging()
  r_version_staging()
}
\seealso{
Other snapshot metadata: 
\code{\link{date_snapshot}()},
\code{\link{date_staging}()}
}
\concept{snapshot metadata}
