% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_includes.R
\name{rclone_includes}
\alias{rclone_includes}
\title{Write Rclone includes.}
\usage{
rclone_includes(path_staging)
}
\arguments{
\item{path_staging}{Character string, directory path to the source
files of the Staging universe.}
}
\description{
Write text files to pass to the \code{--include-from} flag
in Rclone when uploading snapshots.
}
\examples{
\dontrun{
url_staging = "https://github.com/r-multiverse/staging"
path_staging <- tempfile()
path_community <- tempfile()
gert::git_clone(url = url_staging, path = path_staging)
stage_candidates(path_staging = path_staging)
rclone_includes(path_staging)
}
}
\seealso{
Other staging: 
\code{\link{filter_meta}()},
\code{\link{freeze_dependencies}()},
\code{\link{stage_candidates}()}
}
\concept{staging}
