% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_advisories.R
\name{read_advisories}
\alias{read_advisories}
\title{Read R Consortium advisories}
\usage{
read_advisories(timeout = 6e+05, retries = 3L)
}
\arguments{
\item{timeout}{Number of milliseconds until the database download times
out.}

\item{retries}{Number of retries to download the database.}
}
\description{
Read the R Consortium R Advisory database
}
\keyword{internal}
