% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_nonstandard_licenses.R
\name{record_nonstandard_licenses}
\alias{record_nonstandard_licenses}
\title{Record nonstandard licenses}
\usage{
record_nonstandard_licenses(
  path_status = "status.json",
  path_nonstandard_licenses = "nonstandard_licenses.json"
)
}
\arguments{
\item{path_status}{Character string, local path to the \code{status.json} file
of the repository.}

\item{path_nonstandard_licenses}{Character string,
output path to write JSON data with the names
and licenses of packages with non-standard licenses.}
}
\value{
\code{NULL} (invisibly). Called for its side effects.
}
\description{
R-multiverse packages must have valid free and
open-source (FOSS) licenses to protect the intellectual
property rights of the package owners
(c.f. \url{https://en.wikipedia.org/wiki/Free_and_open-source_software}).
\code{\link[=record_nonstandard_licenses]{record_nonstandard_licenses()}} records packages with nonstandard
licenses.
}
