% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_versions.R
\name{record_versions}
\alias{record_versions}
\title{Record the manifest of package versions.}
\usage{
record_versions(
  manifest = "versions.json",
  repos = "https://r-releases.r-universe.dev"
)
}
\arguments{
\item{manifest}{Character of length 1, file path to the JSON manifest.}

\item{repos}{Character string of package repositories to track.}
}
\value{
\code{NULL} (invisibly). Writes a package manifest as a JSON file.
}
\description{
Record the manifest of versions of packages
and their MD5 hashes.
}
\details{
As well as their versions and MD5 hashes, this function
records the highest version ever recorded and the MD5 of that
version. This information helps check if a package complies with
best practices for version numbers: the version number should increment
on each new release.
}
\keyword{internal}
