% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_license.R
\name{review_license}
\alias{review_license}
\title{Review a package license.}
\usage{
review_license(license)
}
\arguments{
\item{license}{Character string with the \code{"License:"} field
of the \code{DESCRIPTION} file of the package in question.}
}
\value{
Invisibly returns
\code{TRUE} if the package has a valid free open-source (FOSS)
license according to \code{\link[tools:licensetools]{tools::analyze_license()}}. \code{FALSE} otherwise.
\code{\link[=review_license]{review_license()}} also prints an R console message
to the communicate the result.
Licenses for which \code{\link[=review_license]{review_license()}} returns \code{FALSE} are
prohibited in R-multiverse.
}
\description{
Review a package license string from the \code{DESCRIPTION} file
to make sure the package has a valid free open-source (FOSS) license.
This is vital to make sure R-multiverse has legal permission
to distribute the package source code.
}
\examples{
  review_license("MIT + file LICENSE")
  review_license("just file LICENSE")
}
\seealso{
Other Manual package reviews: 
\code{\link{review_package}()}
}
\concept{Manual package reviews}
