% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_package.R
\name{review_package}
\alias{review_package}
\title{Review a package.}
\usage{
review_package(name, url, advisories = NULL)
}
\arguments{
\item{name}{Character string, name of the package to check.}

\item{url}{Either a character string with the package URL or
a custom JSON string with a package entry.}

\item{advisories}{Character vector of names of packages with advisories
in the R Consortium Advisory Database.
If \code{NULL}, then \code{review_package_text()} downloads the advisory database and
checks if the package has a vulnerability listed there.
The advisory database is cached internally for performance.}
}
\value{
Invisibly returns \code{TRUE}
if there is a problem with the package entry,
otherwise \code{FALSE} if there are no issues.
In either case, \code{\link[=review_package]{review_package()}} prints an R console message with
the result.

For security reasons, \code{\link[=review_package]{review_package()}} might only print the first
finding it encounters. If that happens, there will be an informative
note at the end of the console message, and compliance with
R-multiverse policies will need to be checked manually.
In particular, please use \code{\link[=review_license]{review_license()}} to
check the \code{"License:"} field in the package \code{DESCRIPTION} file.
}
\description{
Review a package for registration in R-multiverse.
}
\details{
\code{\link[=review_package]{review_package()}} runs all the checks from
\url{https://r-multiverse.org/review.html#automatic-acceptance}
that can be done using the package name and source code repository URL.
}
\examples{
  review_package(
    name = "webchem",
    url = "https://github.com/ropensci/webchem"
  )
  review_package(
    name = "polars",
    url = "https://github.com/pola-rs/r-polars"
  )
}
\seealso{
Other Manual package reviews: 
\code{\link{review_license}()}
}
\concept{Manual package reviews}
