% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_pull_request.R
\name{review_pull_request}
\alias{review_pull_request}
\title{Review a pull request.}
\usage{
review_pull_request(owner = "r-releases", repo = "r-releases", number)
}
\arguments{
\item{owner}{Character of length 1, name of the universe repository owner.}

\item{repo}{Character of length 1, name of the universe repository.}

\item{number}{Positive integer of length 1, index of the pull request
in the repo.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Review a pull request to add packages to \code{packages.json}.
}
\section{Testing}{

Testing of this function unfortunately needs to be manual. Test cases:
\enumerate{
\item Add a package correctly (automatically merge).
\item Add a bad URL (manual review).
\item Change a URL (manual review).
\item Add a file in a forbidden place (close).
\item Add a custom JSON file which can be parsed (manual review).
}
}

\seealso{
Other pull request reviews: 
\code{\link{review_pull_requests}()}
}
\concept{pull request reviews}
