% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_pull_requests.R
\name{review_pull_requests}
\alias{review_pull_requests}
\title{Review pull requests.}
\usage{
review_pull_requests(owner = "r-multiverse", repo = "contributions")
}
\arguments{
\item{owner}{Character of length 1, name of the package repository owner.}

\item{repo}{Character of length 1, URL of the package repository.
R-multiverse uses \code{"https://community.r-multiverse.org"}.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Review pull requests which add packages to \code{packages.json}.
}
\seealso{
Other pull request reviews: 
\code{\link{review_pull_request}()}
}
\concept{pull request reviews}
