% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_is_active.R
\name{staging_is_active}
\alias{staging_is_active}
\title{Check if the stating universe is active.}
\usage{
staging_is_active(
  thresholds = c("01-15", "04-15", "07-15", "10-15"),
  duration = 30L,
  today = format(Sys.Date(), "\%Y-\%m-\%d")
)
}
\arguments{
\item{thresholds}{Character vector of \code{"\%m-\%d"} dates that the
staging universe becomes active.}

\item{duration}{Positive integer, number of days that the staging
universe remains active after each threshold.}

\item{today}{Character string with today's date in \code{"\%Y-\%m-\%d"} format.}
}
\value{
\code{TRUE} if the staging universe is active, \code{FALSE} otherwise.
}
\description{
Check if the stating universe is active.
}
\examples{
staging_is_active()
}
\seealso{
Other staging: 
\code{\link{update_staging}()}
}
\concept{staging}
