% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_is_resetting.R
\name{staging_is_resetting}
\alias{staging_is_resetting}
\title{Check if the stating universe is resetting.}
\usage{
staging_is_resetting(
  start = c("12-15", "03-15", "06-15", "09-15"),
  today = Sys.Date()
)
}
\arguments{
\item{start}{Character vector of \code{"\%m-\%d"} dates that the
staging universe is resetting. The resetting period will then last for a
full calendar month.
For example, if you supply a start date of \code{"12-15"},
then the staging period will include all days from \code{"12-15"}
through \code{"01-14"} and not include \code{"01-15"}.}

\item{today}{Character string with today's date in \code{"\%Y-\%m-\%d"} format or an
object convertible to POSIXlt format.}
}
\value{
\code{TRUE} if the staging universe is resetting, \code{FALSE} otherwise.
}
\description{
Check if the stating universe is resetting.
}
\examples{
staging_is_resetting()
}
\seealso{
Other staging: 
\code{\link{propose_snapshot}()},
\code{\link{staging_is_active}()},
\code{\link{update_staging}()}
}
\concept{staging}
