% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_r_version.R, R/utils_staging.R
\name{staging_r_version}
\alias{staging_r_version}
\title{Staging R version}
\usage{
staging_r_version()

staging_r_version()
}
\value{
A data frame with one row and columns for the \code{POSIXct} date,
patch version, and minor version of the targeted version of R.

A data frame with one row and columns for the \code{POSIXct} date,
patch version, and minor version of the targeted version of R.
}
\description{
Get the version of R from the date of the start of the
most recent Staging period.

Get the version of R from the date of the start of the
most recent Staging period.
}
\examples{
  staging_r_version()
  staging_r_version()
}
