% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_r_version.R, R/staging_start.R
\name{staging_start}
\alias{staging_start}
\title{Staging date}
\usage{
staging_start()

staging_start()
}
\value{
Character string in \code{yyyy-mm-dd} format.

Character string in \code{yyyy-mm-dd} format.
}
\description{
Get the date of the start of the
most recent Staging period.

Get the date of the start of the
most recent Staging period.
}
\examples{
  staging_start()
  staging_start()
}
