% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_checks.R
\name{status_checks}
\alias{status_checks}
\title{Report status from R-universe package check results.}
\usage{
status_checks(meta = meta_checks())
}
\arguments{
\item{meta}{A data frame with R-universe package check results
returned by \code{\link[=meta_checks]{meta_checks()}}.}
}
\value{
A named list of information about packages which do not comply
with \code{DESCRPTION} checks. Each name is a package name,
and each element contains specific information about
non-compliance.
}
\description{
Check R-universe package check results.
}
\details{
\code{\link[=status_checks]{status_checks()}} reads output from
the R-universe check API
to scan all R-multiverse packages for status that may have
happened during building and testing.
}
\section{Package status}{

Functions like \code{\link[=status_versions]{status_versions()}} and \code{\link[=status_descriptions]{status_descriptions()}}
perform health checks for all packages in R-multiverse.
For a complete list of checks, see
the \verb{status_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference/index.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_status]{record_status()}} gathers
together all the status about R-multiverse packages.
}

\examples{
  meta <- meta_checks(repo = "https://wlandau.r-universe.dev")
  status <- status_checks(meta = meta)
  str(status)
}
\seealso{
Other status: 
\code{\link{interpret_status}()},
\code{\link{status_dependencies}()},
\code{\link{status_descriptions}()},
\code{\link{status_versions}()},
\code{\link{update_status}()}
}
\concept{status}
