% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_dependencies.R
\name{status_dependencies}
\alias{status_dependencies}
\title{Report package dependency status}
\usage{
status_dependencies(packages, meta = meta_packages(), verbose = FALSE)
}
\arguments{
\item{packages}{Character vector of names of packages with other issues.}

\item{meta}{A data frame with R-universe package check results
returned by \code{\link[=meta_packages]{meta_packages()}}.}

\item{verbose}{\code{TRUE} to print progress while checking
dependency status, \code{FALSE} otherwise.}
}
\value{
A nested list of problems triggered by dependencies.

To illustrate the structure of this list, suppose
Package \code{tarchetypes} depends on package \code{targets}, and packages
\code{jagstargets} and \code{stantargets} depend on \code{tarchetypes}.
In addition, package \code{targets} has a problem in \verb{R CMD check}
which might cause problems in \code{tarchetypes} and packages downstream.

\code{status_dependencies()} represents this information in the
following list:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  jagstargets = list(targets = "tarchetypes"),
  tarchetypes = list(targets = character(0)),
  stantargets = list(targets = "tarchetypes")
)
}\if{html}{\out{</div>}}

In general, the returned list is of the form:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  impacted_reverse_dependency = list(
    upstream_culprit = c("direct_dependency_1", "direct_dependency_2")
  )
)
}\if{html}{\out{</div>}}

where \code{upstream_culprit} causes problems in \code{impacted_reverse_dependency}
through direct dependencies \code{direct_dependency_1} and
\code{direct_dependency_2}.
}
\description{
Flag packages which have issues in their strong dependencies
(\verb{Imports:}, \verb{Depends:}, and \verb{LinkingTo:} in the \code{DESCRIPTION}.)
These include indirect/upstream dependencies, as well, not just
the explicit mentions in the \code{DESCRIPTION} file.
}
\section{Package status}{

Functions like \code{\link[=status_versions]{status_versions()}} and \code{\link[=status_descriptions]{status_descriptions()}}
perform health checks for all packages in R-multiverse.
For a complete list of checks, see
the \verb{status_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference/index.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_status]{record_status()}} gathers
together all the status about R-multiverse packages.
}

\examples{
  meta <- meta_packages(repo = "https://wlandau.r-universe.dev")
  status_dependencies(packages = character(0L), meta = meta)
  status_dependencies(packages = "crew.aws.batch", meta = meta)
  status_dependencies(packages = "nanonext", meta = meta)
  status_dependencies(packages = "crew", meta = meta)
  status_dependencies(packages = c("crew", "mirai"), meta = meta)
}
\seealso{
Other status: 
\code{\link{interpret_status}()},
\code{\link{status_descriptions}()},
\code{\link{status_r_cmd_check}()},
\code{\link{status_versions}()},
\code{\link{update_status}()}
}
\concept{status}
