% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_descriptions.R
\name{status_descriptions}
\alias{status_descriptions}
\title{Report \code{DESCRIPTION}-level status.}
\usage{
status_descriptions(meta = meta_packages())
}
\arguments{
\item{meta}{A data frame with R-universe package check results
returned by \code{\link[=meta_packages]{meta_packages()}}.}
}
\value{
A named list of information about packages which do not comply
with \code{DESCRPTION} checks. Each name is a package name,
and each element contains specific information about
non-compliance.
}
\description{
Report issues with \code{DESCRIPTION}-level metadata of packages.
}
\details{
\code{\link[=status_descriptions]{status_descriptions()}} reports specific issues in the
\code{DESCRIPTION}-level metadata of packages:
\enumerate{
\item Security advisories at
\url{https://github.com/RConsortium/r-advisory-database}.
\item Licenses that cannot be verified as free and open-source.
\item The presence of a \code{"Remotes"} field.
\item A lower version number on R-multiverse than on CRAN,
if the package is also published there.
}
}
\section{Package status}{

Functions like \code{\link[=status_versions]{status_versions()}} and \code{\link[=status_descriptions]{status_descriptions()}}
perform health checks for all packages in R-multiverse.
For a complete list of checks, see
the \verb{status_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference/index.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_status]{record_status()}} gathers
together all the status about R-multiverse packages.
}

\examples{
  meta <- meta_packages(repo = "https://wlandau.r-universe.dev")
  status <- status_descriptions(meta = meta)
  str(status)
}
\seealso{
Other status: 
\code{\link{interpret_status}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_r_cmd_check}()},
\code{\link{issues_versions}()},
\code{\link{update_status}()}
}
\concept{status}
