% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_versions.R
\name{status_versions}
\alias{status_versions}
\title{Check package versions.}
\usage{
status_versions(versions)
}
\arguments{
\item{versions}{Character of length 1, file path to a JSON manifest
tracking the history of released versions of packages.}
}
\value{
A named list of information about packages which do not comply
with version number history checks. Each name is a package name,
and each element contains specific information about version
non-compliance: the current version number, the current version hash,
and the analogous versions and hashes of the highest-versioned
release recorded.
}
\description{
Check package version number history for compliance.
}
\details{
This function checks the version number history of packages
in R-multiverse and reports any packages with issues. The current
released version of a given package must be unique, and it must be
greater than all the versions of all the previous package releases.
}
\section{Package status}{

Functions like \code{\link[=status_versions]{status_versions()}} and \code{\link[=status_descriptions]{status_descriptions()}}
perform health checks for all packages in R-multiverse.
For a complete list of checks, see
the \verb{status_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference/index.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_status]{record_status()}} gathers
together all the status about R-multiverse packages.
}

\examples{
  lines <- c(
    "[",
    " {",
    " \"package\": \"package_unmodified\",",
    " \"version_current\": \"1.0.0\",",
    " \"hash_current\": \"hash_1.0.0\",",
    " \"version_highest\": \"1.0.0\",",
    " \"hash_highest\": \"hash_1.0.0\"",
    " },",
    " {",
    " \"package\": \"version_decremented\",",
    " \"version_current\": \"0.0.1\",",
    " \"hash_current\": \"hash_0.0.1\",",
    " \"version_highest\": \"1.0.0\",",
    " \"hash_highest\": \"hash_1.0.0\"",
    " },",
    " {",
    " \"package\": \"version_incremented\",",
    " \"version_current\": \"2.0.0\",",
    " \"hash_current\": \"hash_2.0.0\",",
    " \"version_highest\": \"2.0.0\",",
    " \"hash_highest\": \"hash_2.0.0\"",
    " },",
    " {",
    " \"package\": \"version_unmodified\",",
    " \"version_current\": \"1.0.0\",",
    " \"hash_current\": \"hash_1.0.0-modified\",",
    " \"version_highest\": \"1.0.0\",",
    " \"hash_highest\": \"hash_1.0.0\"",
    " }",
    "]"
  )
  versions <- tempfile()
  writeLines(lines, versions)
  out <- status_versions(versions)
  str(out)
}
\seealso{
Other status: 
\code{\link{interpret_status}()},
\code{\link{status_checks}()},
\code{\link{status_dependencies}()},
\code{\link{status_descriptions}()},
\code{\link{update_status}()}
}
\concept{status}
