% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{try_message}
\alias{try_message}
\title{Get try error message}
\usage{
try_message(try_error, collapse = " ")
}
\arguments{
\item{try_error}{\code{try()} error object.}

\item{collapse}{Character of length 1, delimiter for the
lines of the message.}
}
\description{
Get the error message of an error object from \code{try()}.
}
\keyword{internal}
