% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_staging.R
\name{update_staging}
\alias{update_staging}
\title{Update staging}
\usage{
update_staging(
  path_staging,
  path_community,
  repo_community = "https://community.r-multiverse.org",
  mock = NULL
)
}
\arguments{
\item{path_staging}{Character string, directory path to the source
files of the staging universe.}

\item{path_community}{Character string, directory path to the source
files of the community universe.}

\item{repo_community}{Character string, URL of the community universe.}

\item{mock}{For testing purposes only, a named list of data frames
for inputs to various intermediate functions.}
}
\value{
\code{NULL} (invisibly)
}
\description{
Update the staging universe.
}
\details{
\code{\link[=update_staging]{update_staging()}} controls how packages enter and leave
the staging universe. It updates the staging \code{packages.json}
manifest depending on the contents of the community
universe and issues with package checks.
}
\examples{
\dontrun{
url_staging = "https://github.com/r-multiverse/staging"
url_community = "https://github.com/r-multiverse/community"
path_staging <- tempfile()
path_community <- tempfile()
gert::git_clone(url = url_staging, path = path_staging)
gert::git_clone(url = url_community, path = path_community)
update_staging(
  path_staging = path_staging,
  path_community = path_community,
  repo_community = "https://community.r-multiverse.org"
)
}
}
