% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_status.R
\name{update_status}
\alias{update_status}
\title{Update the package status repository}
\usage{
update_status(
  path_status,
  path_staging,
  path_community,
  repo_staging = "https://staging.r-multiverse.org",
  repo_community = "https://community.r-multiverse.org",
  mock = NULL
)
}
\arguments{
\item{path_status}{Character string, directory path to the source files
of the package status repository.}

\item{path_staging}{Character string, directory path to the source
files of the staging universe.}

\item{path_community}{Character string, directory path to the source
files of the community universe.}

\item{repo_community}{Character string, URL of the community universe.}

\item{mock}{For testing purposes only, a named list of data frames
for inputs to various intermediate functions.}
}
\description{
Update the repository which reports the status on individual
packages.
}
\examples{
\dontrun{
url_staging = "https://github.com/r-multiverse/staging"
url_community = "https://github.com/r-multiverse/community"
path_status <- tempfile()
path_staging <- tempfile()
path_community <- tempfile()
gert::git_clone(url = url_staging, path = path_staging)
gert::git_clone(url = url_community, path = path_community)
update_status(
  path_status = path_status,
  path_staging = path_staging,
  path_community = path_community,
  repo_staging = "https://staging.r-multiverse.org",
  repo_community = "https://community.r-multiverse.org"
)
}
}
\seealso{
Other status: 
\code{\link{interpret_status}()}
}
\concept{status}
