% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_topics.R
\name{update_topics}
\alias{update_topics}
\title{Update topics}
\usage{
update_topics(path, repo = "https://community.r-multiverse.org", mock = NULL)
}
\arguments{
\item{path}{Character string,
local file path to the topics repository source code.}

\item{repo}{Character string, URL of the Community universe.}

\item{mock}{For testing purposes only, a named list of data frames
for inputs to various intermediate functions.}
}
\value{
\code{NULL} (invisibly). Called for its side effects.
}
\description{
Update the list of packages for each
R-multiverse topic.
}
\concept{topics}
