% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_package.R
\name{assert_release_exists}
\alias{assert_release_exists}
\title{Check for a release.}
\usage{
assert_release_exists(url)
}
\arguments{
\item{url}{Usually a character of length 1 with the package URL.}
}
\value{
A character string if there is a problem with the package entry,
otherwise \code{NULL} if there are no issues.
}
\description{
Check for a release.
}
\keyword{internal}
