% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_universe.R
\name{build_universe}
\alias{build_universe}
\title{Build the universe.}
\usage{
build_universe(input = getwd(), output = "packages.json")
}
\arguments{
\item{input}{Character of length 1, directory path with the
text file listings of R releases.}

\item{output}{Character of length 1, file path where the
\code{r-universe} \code{packages.json} file will be written.}
}
\value{
NULL (invisibly)
}
\description{
Create the \code{r-universe} \code{packages.json} file
from constituent text files with URLs.
}
\keyword{internal}
