% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_descriptions.R
\name{check_descriptions}
\alias{check_descriptions}
\title{Check package \code{DESCRIPTION} files.}
\usage{
check_descriptions(repo = "https://multiverse.r-multiverse.org", mock = NULL)
}
\arguments{
\item{repo}{Character of length 1, URL of the repository
of R package candidates for production.}

\item{mock}{For testing purposes only,
an optional pre-computed data frame with details about packages
for this type of check.}
}
\value{
A named list of information about packages which do not comply
with \code{DESCRPTION} checks. Each name is a package name,
and each element contains specific information about
non-compliance.
}
\description{
Check the \code{DESCRIPTION} files of packages for specific
issues.
}
\details{
This function scrapes the \code{src/contrib/PACKAGES.json} file
of the universe to check the data in the \code{DESCRIPTION} files of packages
for compliance. Right now, the only field checked is \verb{Remotes:}.
A packages with a \verb{Remotes:} field in the \code{DESCRIPTION} file may
depend on development versions of other packages and so are
excluded from the production universe.
}
\section{Package checks for production}{

Functions like \code{\link[=check_versions]{check_versions()}} and \code{\link[=check_descriptions]{check_descriptions()}}
perform health checks for all packages in R-multiverse.
Only packages that pass these checks go to the production repository at
\url{https://production.r-multiverse.org}. For a complete list of checks, see
the \verb{check_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_issues]{record_issues()}} gathers
together all the issues about R-multiverse packages.
}

\examples{
  str(check_descriptions(repo = "https://multiverse.r-multiverse.org"))
}
\seealso{
Other package checks for production: 
\code{\link{check_checks}()},
\code{\link{check_versions}()}
}
\concept{package checks for production}
