% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_packages.R
\name{filter_packages}
\alias{filter_packages}
\title{Filter \code{PACKAGES} and \code{PACKAGES.gz} files.}
\usage{
filter_packages(path_packages, path_staging)
}
\arguments{
\item{path_packages}{Directory path where \code{PACKAGES} and \code{PACKAGES.gz} files
reside locally.}

\item{path_staging}{Path to a GitHub clone of the Staging universe.}
}
\description{
Filter the \code{PACKAGES} and \code{PACKAGES.gz} files to
only list certain packages.
}
\examples{
\dontrun{
path_packages <- tempfile()
dir.create(path_packages)
mock <- system.file(
  "packages",
  package = "multiverse.internals",
  mustWork = TRUE
)
file.copy(mock, path_packages, recursive = TRUE)
path_staging <- tempfile()
url_staging <- "https://github.com/r-multiverse/staging"
gert::git_clone(url = url_staging, path = path_staging)
filter_packages(path_packages, path_staging)
}
}
\seealso{
Other staging: 
\code{\link{freeze_dependencies}()},
\code{\link{rclone_includes}()},
\code{\link{stage_candidates}()}
}
\concept{staging}
