% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_status.R
\name{interpret_status}
\alias{interpret_status}
\title{Interpret the status of a package}
\usage{
interpret_status(package, issues, is_staging)
}
\arguments{
\item{package}{Character string, name of the package.}

\item{issues}{A list with one issue per package. Obtained by
reading the results of \code{\link[=record_issues]{record_issues()}}.}
}
\value{
A character string summarizing the issues of a package in prose.
}
\description{
Summarize the issues of a package in human-readable text.
}
\seealso{
Other status: 
\code{\link{update_status}()}
}
\concept{status}
