% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_descriptions.R
\name{issues_descriptions}
\alias{issues_descriptions}
\title{Check package \code{DESCRIPTION} files.}
\usage{
issues_descriptions(meta)
}
\arguments{
\item{meta}{A data frame of package metadata returned by \code{\link[=meta_packages]{meta_packages()}}.}
}
\value{
A named list of information about packages which do not comply
with \code{DESCRPTION} checks. Each name is a package name,
and each element contains specific information about
non-compliance.
}
\description{
Check the \code{DESCRIPTION} files of packages for specific
issues.
}
\details{
This function scrapes the \code{src/contrib/PACKAGES.json} file
of the universe to check the data in the \code{DESCRIPTION} files of packages
for compliance. Right now, the only field checked is \verb{Remotes:}.
A packages with a \verb{Remotes:} field in the \code{DESCRIPTION} file may
depend on development versions of other packages and so are
excluded from the production universe.
}
\section{Package issues}{

Functions like \code{\link[=issues_versions]{issues_versions()}} and \code{\link[=issues_descriptions]{issues_descriptions()}}
perform health checks for all packages in R-multiverse.
Only packages that pass these checks go to the production repository at
\url{https://production.r-multiverse.org}. For a complete list of checks, see
the \verb{issues_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_issues]{record_issues()}} gathers
together all the issues about R-multiverse packages.
}

\examples{
  meta <- meta_packages(repo = "https://wlandau.r-universe.dev")
  issues <- issues_descriptions(meta = meta)
  str(issues)
}
\seealso{
Other issues: 
\code{\link{issues_checks}()},
\code{\link{issues_versions}()}
}
\concept{issues}
