% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_descriptions.R
\name{issues_descriptions}
\alias{issues_descriptions}
\title{Report \code{DESCRIPTION} file issues.}
\usage{
issues_descriptions(meta = meta_packages())
}
\arguments{
\item{meta}{A data frame with R-universe package check results
returned by \code{\link[=meta_checks]{meta_checks()}}.}
}
\value{
A named list of information about packages which do not comply
with \code{DESCRPTION} checks. Each name is a package name,
and each element contains specific information about
non-compliance.
}
\description{
Report issues with the \code{DESCRIPTION} files of packages.
}
\details{
\code{\link[=issues_descriptions]{issues_descriptions()}} scans downloaded metadata from the
\code{PACKAGES.json} file of an R universe and reports issues with a
package's description file, such as the presence of a
\code{"Remotes"} field.
}
\section{Package issues}{

Functions like \code{\link[=issues_versions]{issues_versions()}} and \code{\link[=issues_descriptions]{issues_descriptions()}}
perform health checks for all packages in R-multiverse.
Only packages that pass these checks go to the production repository at
\url{https://production.r-multiverse.org}. For a complete list of checks, see
the \verb{issues_*()} functions listed at
\url{https://r-multiverse.org/multiverse.internals/reference.html}.
\code{\link[=record_versions]{record_versions()}} updates the version number history
of releases in R-multiverse, and \code{\link[=record_issues]{record_issues()}} gathers
together all the issues about R-multiverse packages.
}

\examples{
  meta <- meta_packages(repo = "https://wlandau.r-universe.dev")
  issues <- issues_descriptions(meta = meta)
  str(issues)
}
\seealso{
Other issues: 
\code{\link{issues_checks}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_versions}()}
}
\concept{issues}
