% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues_versions.R
\name{issues_versions}
\alias{issues_versions}
\title{Package version issues.}
\usage{
issues_versions(versions)
}
\arguments{
\item{versions}{Character of length 1, file path to a JSON manifest
tracking the history of released versions of packages.}
}
\value{
A data frame with one row for each problematic
package and columns with
the package names and version issues.
}
\description{
Check package version number history for compliance.
}
\details{
This function checks the version number history of packages
in R-multiverse and reports any packages with issues. The current
released version of a given package must be unique, and it must be
greater than all the versions of all the previous package releases.
}
\examples{
  lines <- c(
    "[",
    " {",
    " \"package\": \"package_unmodified\",",
    " \"version_current\": \"1.0.0\",",
    " \"hash_current\": \"hash_1.0.0\",",
    " \"version_highest\": \"1.0.0\",",
    " \"hash_highest\": \"hash_1.0.0\"",
    " },",
    " {",
    " \"package\": \"version_decremented\",",
    " \"version_current\": \"0.0.1\",",
    " \"hash_current\": \"hash_0.0.1\",",
    " \"version_highest\": \"1.0.0\",",
    " \"hash_highest\": \"hash_1.0.0\"",
    " },",
    " {",
    " \"package\": \"version_incremented\",",
    " \"version_current\": \"2.0.0\",",
    " \"hash_current\": \"hash_2.0.0\",",
    " \"version_highest\": \"2.0.0\",",
    " \"hash_highest\": \"hash_2.0.0\"",
    " },",
    " {",
    " \"package\": \"version_unmodified\",",
    " \"version_current\": \"1.0.0\",",
    " \"hash_current\": \"hash_1.0.0-modified\",",
    " \"version_highest\": \"1.0.0\",",
    " \"hash_highest\": \"hash_1.0.0\"",
    " }",
    "]"
  )
  versions <- tempfile()
  writeLines(lines, versions)
  out <- issues_versions(versions)
  str(out)
}
\seealso{
Other issues: 
\code{\link{issues_advisories}()},
\code{\link{issues_dependencies}()},
\code{\link{issues_licenses}()},
\code{\link{issues_r_cmd_check}()},
\code{\link{issues_remotes}()},
\code{\link{issues_version_conflicts}()}
}
\concept{issues}
