% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_includes.R
\name{rclone_includes}
\alias{rclone_includes}
\title{Write Rclone includes.}
\usage{
rclone_includes(
  path_staging,
  repo_staging = "https://staging.r-multiverse.org",
  mock = NULL
)
}
\arguments{
\item{path_staging}{Character string, directory path to the source
files of the Staging universe.}

\item{repo_staging}{Character string, URL of the Staging universe.}

\item{mock}{For testing purposes only, a named list of data frames
for inputs to various intermediate functions.}
}
\description{
Write text files to pass to the \code{--include-from} flag
in Rclone when uploading snapshots.
}
\examples{
\dontrun{
url_staging = "https://github.com/r-multiverse/staging"
url_community = "https://github.com/r-multiverse/community"
path_staging <- tempfile()
path_community <- tempfile()
gert::git_clone(url = url_staging, path = path_staging)
gert::git_clone(url = url_community, path = path_community)
stage_candidates(
  path_staging = path_staging,
  path_community = path_community,
  repo_community = "https://community.r-multiverse.org"
)
rclone_includes(
  path_staging,
  repo_staging = "https://staging.r-multiverse.org"
)
}
}
\seealso{
Other staging: 
\code{\link{filter_packages}()},
\code{\link{freeze_dependencies}()},
\code{\link{stage_candidates}()}
}
\concept{staging}
