% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_versions.R
\name{record_versions}
\alias{record_versions}
\title{Record the manifest of package versions.}
\usage{
record_versions(
  versions = "versions.json",
  repo = "https://multiverse.r-multiverse.org",
  current = multiverse.internals::get_current_versions(repo = repo)
)
}
\arguments{
\item{versions}{Character of length 1, file path to a JSON manifest
tracking the history of released versions of packages.
The official versions file for R-multiverse is maintained and
updated periodically at
\url{https://github.com/r-multiverse/checks/blob/main/versions.json}.}

\item{repo}{Character string of package repositories to track.}

\item{current}{A data frame of current versions and hashes of packages
in \code{repo}. This argument is exposed for testing only.}
}
\value{
\code{NULL} (invisibly). Writes a package version manifest
and a manifest of version issues as JSON files.
}
\description{
Record the manifest of versions of packages
and their hashes.
}
\details{
This function tracks a manifest containing the current version,
the current hash, the highest version ever released, and
the hash of the highest version ever released. It uses this information
to determine whether the package complies with best
practices for version numbers.
}
\examples{
  # R-multiverse uses https://multiverse.r-multiverse.org as the repo.
  repo <- "https://wlandau.r-universe.dev" # just for testing and examples
  output <- tempfile()
  versions <- tempfile()
  # First snapshot:
  record_versions(
    versions = versions,
    repo = repo
  )
  readLines(versions)
  # In subsequent snapshots, we have historical information about versions.
  record_versions(
    versions = versions,
    repo = repo
  )
  readLines(versions)
}
