% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_versions.R
\name{record_versions}
\alias{record_versions}
\title{Record the manifest of package versions.}
\usage{
record_versions(
  manifest = "versions.json",
  issues = "version_issues.json",
  repos = "https://r-releases.r-universe.dev",
  current = r.releases.utils::get_current_versions(repos = repos)
)
}
\arguments{
\item{manifest}{Character of length 1, file path to the JSON manifest.}

\item{issues}{Character of length 1, file path to a JSON file
which records packages with version issues.}

\item{repos}{Character string of package repositories to track.}

\item{current}{A data frame of current versions and hashes of packages
in \code{repos}. This argument is exposed for testing only.}
}
\value{
\code{NULL} (invisibly). Writes a package version manifest
and a manifest of version issues as JSON files.
}
\description{
Record the manifest of versions of packages
and their MD5 hashes.
}
\details{
This function tracks a manifest containing the current version,
the current MD5 hash, the highest version ever released, and
the MD5 hash of the highest version ever released. Each time it runs,
it reads scrapes the package repository for the current releases,
reads the old manifest, and updates recorded highest version and MD5
for all packages which incremented their version numbers.
After recording these incremented versions, the current version should
be the highest version, and the current and highest-version
MD5 hashes should agree. Packages
that fall out of alignment are recorded in a small JSON with only
the packages with version issues.
}
