% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_versions.R
\name{record_versions}
\alias{record_versions}
\title{Record the manifest of package versions.}
\usage{
record_versions(
  manifest = "versions.json",
  repo = "https://r-multiverse.r-universe.dev",
  current = multiverse.internals::get_current_versions(repo = repo)
)
}
\arguments{
\item{manifest}{Character of length 1, file path to the JSON manifest.}

\item{repo}{Character string of package repositories to track.}

\item{current}{A data frame of current versions and hashes of packages
in \code{repo}. This argument is exposed for testing only.}
}
\value{
\code{NULL} (invisibly). Writes a package version manifest
and a manifest of version issues as JSON files.
}
\description{
Record the manifest of versions of packages
and their hashes.
}
\details{
This function tracks a manifest containing the current version,
the current hash, the highest version ever released, and
the hash of the highest version ever released. It uses this information
to determine whether the package complies with best
practices for version numbers.
}
