% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_package.R
\name{review_package}
\alias{review_package}
\title{Review a package.}
\usage{
review_package(name, url, advisories = NULL)
}
\arguments{
\item{name}{Character of length 1, package name.}

\item{url}{Usually a character of length 1 with the package URL.}

\item{advisories}{Character vector of names of packages with advisories
in the R Consortium Advisory Database.
If \code{NULL}, then \code{review_package()} downloads the advisory database and
checks if the package has a vulnerability listed there.
The advisory database is cached internally for performance.}
}
\value{
A character string if there is a problem with the package entry,
otherwise \code{NULL} if there are no issues.
For security reasons, \code{\link[=review_package]{review_package()}} might only return the first
finding it encounters. If that happens, there will be an informative
note at the end of the text string.
}
\description{
Review a package for registration in R-multiverse.
}
\details{
\code{\link[=review_package]{review_package()}} runs all the checks from
\url{https://r-multiverse.org/review.html#automatic-acceptance}
that can be done using the package name and source code repository URL.
}
\examples{
  review_package(
    name = "multiverse.internals",
    url = "https://github.com/r-multiverse/multiverse.internals"
  )
}
\seealso{
Other package reviews: 
\code{\link{review_license}()},
\code{\link{review_pull_request}()},
\code{\link{review_pull_requests}()}
}
\concept{package reviews}
