% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_pull_request.R
\name{review_pull_request}
\alias{review_pull_request}
\title{Review an R-multiverse contribution}
\usage{
review_pull_request(
  owner = "r-multiverse",
  repo = "contributions",
  number,
  advisories = character(0L)
)
}
\arguments{
\item{owner}{Character of length 1, name of the package repository owner.}

\item{repo}{Character of length 1, URL of the package repository.
R-multiverse uses \code{"https://community.r-multiverse.org"}.}

\item{number}{Positive integer of length 1, index of the pull request
in the repo.}

\item{advisories}{Character vector of names of packages with advisories
in the R Consortium Advisory Database.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Review a pull request to add packages to R-multiverse.
}
\seealso{
Other pull request reviews: 
\code{\link{review_pull_requests}()}
}
\concept{pull request reviews}
