% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_pull_requests.R
\name{review_pull_requests}
\alias{review_pull_requests}
\title{Review R-multiverse contribution pull requests.}
\usage{
review_pull_requests(owner = "r-multiverse", repo = "contributions")
}
\arguments{
\item{owner}{Character of length 1, name of the package repository owner.}

\item{repo}{URL of the repository to query.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Review pull requests which add packages to \code{packages.json}.
}
\seealso{
Other package reviews: 
\code{\link{review_license}()},
\code{\link{review_package}()},
\code{\link{review_pull_request}()}
}
\concept{package reviews}
