% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_pull_requests.R
\name{review_pull_requests}
\alias{review_pull_requests}
\title{Review pull requests.}
\usage{
review_pull_requests(owner = "r-multiverse", repo = "contributions")
}
\arguments{
\item{owner}{Character of length 1, name of the package repository owner.}

\item{repo}{Character of length 1, URL of the package repository.
R-multiverse uses \code{"https://multiverse.r-multiverse.org"}.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Review pull requests which add packages to \code{packages.json}.
}
\section{Testing}{

Testing of this function unfortunately needs to be manual. Test cases:
\enumerate{
\item Add a package correctly (automatically merge).
\item Add a bad URL (manual review).
\item Change a URL (manual review).
\item Add a file in a forbidden place (close).
\item Add a custom JSON file which can be parsed (manual review).
}
}

\seealso{
Other pull request reviews: 
\code{\link{review_pull_request}()}
}
\concept{pull request reviews}
