% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_candidates.R
\name{stage_candidates}
\alias{stage_candidates}
\title{Stage release candidates}
\usage{
stage_candidates(
  path_staging,
  path_community,
  repo_community = "https://community.r-multiverse.org",
  types = c("src", "win", "mac"),
  mock = NULL
)
}
\arguments{
\item{path_staging}{Character string, directory path to the source
files of the staging universe.}

\item{path_community}{Character string, directory path to the source
files of the community universe.}

\item{repo_community}{Character string, URL of the community universe.}

\item{types}{Character vector, what to pass to the \code{types} field in the
snapshot API URL. Controls the types of binaries and documentation
included in the snapshot.}

\item{mock}{For testing purposes only, a named list of data frames
for inputs to various intermediate functions.}
}
\value{
\code{NULL} (invisibly)
}
\description{
Stage release candidates for the targeted Production snapshot.
}
\details{
\code{\link[=stage_candidates]{stage_candidates()}} Writes \code{packages.json} to control
contents of the Staging universe.
It also writes \code{staged.json} to track packages staged for Production
and \code{snapshot.url} with an API call to download staged packages.
}
\examples{
\dontrun{
url_staging = "https://github.com/r-multiverse/staging"
url_community = "https://github.com/r-multiverse/community"
path_staging <- tempfile()
path_community <- tempfile()
gert::git_clone(url = url_staging, path = path_staging)
gert::git_clone(url = url_community, path = path_community)
stage_candidates(
  path_staging = path_staging,
  path_community = path_community,
  repo_community = "https://community.r-multiverse.org"
)
}
}
\concept{staging}
