% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_is_active.R
\name{staging_is_active}
\alias{staging_is_active}
\title{Check if the stating universe is active.}
\usage{
staging_is_active(
  start = c("01-15", "04-15", "07-15", "10-15"),
  today = Sys.Date()
)
}
\arguments{
\item{start}{Character vector of \code{"\%m-\%d"} dates that the
staging universe becomes active. Staging will then last for a full
calendar month. For example, if you supply a start date of \code{"01-15"},
then the staging period will include all days from \code{"01-15"}
through \code{"02-14"} and not include \code{"02-15"}.}

\item{today}{Character string with today's date in \code{"\%Y-\%m-\%d"} format or an
object convertible to POSIXlt format.}
}
\value{
\code{TRUE} if the staging universe is active, \code{FALSE} otherwise.
}
\description{
Check if the stating universe is active.
}
\examples{
staging_is_active()
}
\seealso{
Other staging: 
\code{\link{propose_snapshot}()},
\code{\link{update_staging}()}
}
\concept{staging}
