% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_topics.R
\name{update_topics}
\alias{update_topics}
\title{Update topics}
\usage{
update_topics(path, repo = "https://community.r-multiverse.org", mock = NULL)
}
\arguments{
\item{path}{Character string,
local file path to the topics repository source code.}

\item{repo}{Character string, URL of the Community universe.}

\item{mock}{List of named objects for testing purposes only.}
}
\value{
\code{NULL} (invisibly). Called for its side effects.
}
\description{
Update the list of packages for each
R-multiverse topic.
}
\examples{
\dontrun{
path <- tempfile()
gert::git_clone("https://github.com/r-multiverse/topics", path = path)
update_topics(path = path, repo = "https://community.r-multiverse.org")
}
}
\concept{topics}
